Move routing into more fitting apiV2 folder
