--- packages/meditrak-server/src/app/index.js
@@ -1,6 +1,0 @@
-/**
- * Tupaia MediTrak
- * Copyright (c) 2017 Beyond Essential Systems Pty Ltd
- */
-
-export { createApp } from './createApp';
